ISSUE #1746: Re-arrange javascript files
