+++ public/js/components/data-tables.js
@@ -1,4 +1,6 @@
-import {ClusterRenderer, DataTableStorage, FilterManager, Sorter} from "../filters";
+import {ClusterRenderer} from "../data-table/cluster-renderer";
+import {FilterManager} from "../data-table/filter-manager";
+import {Sorter} from "../data-table/sorter";
 import {debounce} from "../utils";
 
 export default class DataTableManager {
