+++ public/js/components/fullscreen.js
