+++ public/js/components/tabs.js
