+++ public/js/data-table/filter-manager.js
@@ -1,6 +1,3 @@
-import {numberFormat} from "./utils";
-import Clusterize from '../libraries/clusterize/clusterize.min';
-
 export class FilterManager {
     constructor(wrapper, storage) {
         this.wrapper = wrapper;
@@ -158,114 +155,5 @@ export class FilterManager {
 
     updateStorage(tableName, activeFilters) {
         this.storage.set(tableName, activeFilters);
-    }
-}
-
-export class SummableCalculator {
-    static calculate(rows) {
-        return rows
-            .filter(r => r.active)
-            .reduce((acc, r) => {
-                Object.entries(r.summables).forEach(([k, v]) => acc[k] = (acc[k] || 0) + v);
-                return acc;
-            }, {});
-    }
-
-    static render(wrapper, rows) {
-        const sums = this.calculate(rows);
-        wrapper.querySelectorAll('[data-dataTable-summable]').forEach(node => {
-            const key = node.getAttribute('data-dataTable-summable');
-            node.innerHTML = sums[key] !== undefined ? numberFormat(sums[key], 0, ',', ' ') : 0;
-        });
-    }
-}
-
-export class Sorter {
-    constructor(columns) {
-        this.columns = columns; // NodeList of <th>
-        this.sortAsc = false;
-        this.sortOnPrevious = null;
-    }
-
-    attachListeners(dataRows, cluster, scrollElem) {
-        this.columns.forEach(th => th.addEventListener('click', () => {
-            const sortOn = th.getAttribute('data-dataTable-sort');
-            if (sortOn === this.sortOnPrevious) this.sortAsc = !this.sortAsc;
-            else this.sortAsc = true;
-            this.sortOnPrevious = sortOn;
-
-            this.columns.forEach(c => c.querySelector('.sorting-icon')?.setAttribute('aria-sort', 'none'));
-            th.querySelector('.sorting-icon')?.setAttribute('aria-sort', this.sortAsc ? 'ascending' : 'descending');
-
-            dataRows.sort((a, b) => {
-                const aVal = a.sort[sortOn], bVal = b.sort[sortOn];
-                if (aVal === undefined) return 1;
-                if (bVal === undefined) return -1;
-                if (aVal < bVal) return this.sortAsc ? -1 : 1;
-                if (aVal > bVal) return this.sortAsc ? 1 : -1;
-                return 0;
-            });
-
-            cluster.update(dataRows.filter(r => r.active).map(r => r.markup));
-            scrollElem.scrollTop = 0;
-        }));
-    }
-}
-
-export class ClusterRenderer {
-    constructor(wrapper, tbody, scrollElem) {
-        this.wrapper = wrapper;
-        this.tbody = tbody;
-        this.scrollElem = scrollElem;
-        this.cluster = null;
-    }
-
-    init(dataRows) {
-        this.cluster = new Clusterize({
-            rows: dataRows.filter(r => r.active).map(r => r.markup),
-            scrollElem: this.scrollElem,
-            contentElem: this.tbody,
-            no_data_class: 'clusterize-loading',
-            callbacks: {
-                clusterChanged: () => {
-                    SummableCalculator.render(this.wrapper, dataRows);
-
-                    const resultCount = this.wrapper.querySelector('[data-dataTable-result-count]');
-                    if (resultCount) resultCount.innerText = dataRows.filter(r => r.active).length;
-
-                    document.dispatchEvent(new CustomEvent('dataTableClusterWasChanged', { bubbles: true, cancelable: false, }));
-                }
-            }
-        });
-    }
-
-    update(dataRows) {
-        if (!this.cluster) return;
-        this.cluster.update(dataRows.filter(r => r.active).map(r => r.markup));
-        this.scrollElem.scrollTop = 0;
-    }
-}
-
-export class DataTableStorage {
-    constructor(storageKey = 'dataTableFilters') {
-        this.storageKey = storageKey;
-    }
-
-    clearAll(tableName) {
-        this.set(tableName, {});
-    }
-
-    get(tableName) {
-        const storedJson = localStorage.getItem(this.storageKey);
-        if (!storedJson) return {};
-        const parsed = JSON.parse(storedJson);
-        return parsed[tableName] || {};
-    }
-
-    set(tableName, object) {
-        const storedJson = localStorage.getItem(this.storageKey);
-        const existing = storedJson ? JSON.parse(storedJson) : {};
-        existing[tableName] = object;
-        localStorage.setItem(this.storageKey, JSON.stringify(existing));
     }
 }
